+++ src/components/Layout.jsx
@@ -1,0 +1,12 @@
+import { Outlet } from 'react-router-dom';
+import Navbar from '@/components/Navbar';
+
+const Layout = () => {
+  return (
+    <div className="wrapper">
+      <Navbar />
+      <Outlet />
+    </div>
+  );
+};
+export default Layout;
