+++ src/components/Navbar2.jsx
@@ -1,0 +1,35 @@
+import { useState, useRef} from "react";
+import { useOnClickOutside } from "@/useOnClickOutside";
+
+
+const Navbar2 = () => {
+  const [dropdown, setDropdown] = useState(false);
+  // useEffect(() => {
+  //   document.title = `Current state value: ${dropdown}`;
+  // }, [dropdown]);
+
+  const ref = useRef();
+
+  useOnClickOutside(ref, dropdown, () => setDropdown(false));
+
+  return (
+    <nav>
+      <ul>
+        <li>Home</li>
+        <li>About</li>
+        <li ref={ref}>
+          <button onClick={() => setDropdown((prev) => !prev)}>
+            Services <span>&#8595;</span>
+          </button>
+          {dropdown && (
+            <ul>
+              <li>Design</li>
+              <li>Development</li>
+            </ul>
+          )}
+        </li>
+      </ul>
+    </nav>
+  );
+};
+export default Navbar2;
