+++ src/components/TodoApp.jsx
@@ -5,15 +5,18 @@ import About from '@/routes/About';
 import Login from '@/routes/Login';
 import Profile from '@/routes/Profile';
 import NotMatch from '@/routes/NoMatch';
+import Layout from '@/components/Layout';
 
 const TodoApp = () => {
   return (
     <Routes>
-      <Route path="/" element={<Home />} />
-      <Route path="about" element={<About />} />
-      <Route path="login" element={<Login />} />
-      <Route path="profile" element={<Profile />} />
-      <Route path="*" element={<NotMatch />} />
+      <Route path="/" element={<Layout />}>
+        <Route index element={<Home />} />
+        <Route path="about" element={<About />} />
+        <Route path="login" element={<Login />} />
+        <Route path="profile" element={<Profile />} />
+        <Route path="*" element={<NotMatch />} />
+      </Route>
     </Routes>
   );
 };
