+++ src/routes/Home.jsx
@@ -3,12 +3,12 @@ import TodosLogic from '@/components/TodosLogic';
 
 const Home = () => {
   return (
-    <div className="wrapper">
+    // <div className="wrapper">
       <div className="todos">
         <Header />
         <TodosLogic />
       </div>
-    </div>
+    // </div>
   );
 };
 export default Home;
