+++ src/styles/app.css
@@ -58,3 +58,12 @@ body {
   margin-top: 5px;
   display: block;
 }
+
+a {
+  color: #333;
+  text-decoration: none;
+}
+/* Active route */
+.active {
+  color: #8a4baf;
+}
